+++ es-core/src/components/ComponentList.cpp
@@ -188,10 +188,10 @@ void ComponentList::render(const Eigen::Affine3f& parentTrans)
 	// draw our entries
 	std::vector<GuiComponent*> drawAfterCursor;
 	bool drawAll;
-	for (unsigned int i = 0; i < mEntries.size(); i++)
+	for (size_t i = 0; i < mEntries.size(); i++)
 	{
 		auto& entry = mEntries.at(i);
-		drawAll = !mFocused || i != mCursor;
+		drawAll = !mFocused || (static_cast<int>(i) != mCursor);
 		for (auto it = entry.data.elements.begin(); it != entry.data.elements.end(); it++)
 		{
 			if (drawAll || it->invert_when_selected)
